bug fix for logging
value
Out[35]: tensor(3211., device='cuda:0')
value.mean() * batch_size  / batch_size
Out[36]: tensor(3210.9998, device='cuda:0')
value.type(torch.float64).mean() * batch_size / batch_size
Out[37]: tensor(3211., device='cuda:0', dtype=torch.float64)

+++ pytorch_lightning/trainer/connectors/logger_connector/result.py
@@ -207,9 +207,9 @@ class ResultMetric(Metric, DeviceDtypeModuleMixin):
         self.meta = metadata
         self.has_reset = False
         if is_tensor:
-            self.add_state("value", torch.tensor(0, dtype=torch.float), dist_reduce_fx=torch.sum)
+            self.add_state("value", torch.tensor(0, dtype=torch.float64), dist_reduce_fx=torch.sum)
             if self.meta.is_mean_reduction:
-                self.add_state("cumulated_batch_size", torch.tensor(0, dtype=torch.float), dist_reduce_fx=torch.sum)
+                self.add_state("cumulated_batch_size", torch.tensor(0, dtype=torch.float64), dist_reduce_fx=torch.sum)
 
     def update(self, value: _IN_METRIC, batch_size: torch.Tensor) -> None:
         if self.is_tensor:
